Update patch set 3
Patch Set 3:

Can we do what was done with ADW in CM7? Don't show this at all if you are running CM, and show an option in the Settings app that links to it. If it's a non-CM/standalone build then it can show the option in the main menu.

Patch-set: 3