Update patch set 4
Patch Set 4:

I will be writing a new drawer system which will allow for different drawer types (vertical), folders in drawer, renaming of apps and that will be on of the features. 

It will use a layering system of the apps list: packages -> drawer database (renames, folders) -> filtered list (sorting, system/user/sdcard app) -> generic drawer implementation.

This will make new drawers much easier to write. The tabs at the top will be in all of the drawers but there will be a "mini" mode to take up less space.

Patch-set: 4