Update patch set 1
Patch Set 1:

I think that explicit imports are better code-wise as it prevents accidental pollution of the namespace (accidentally using another class of the same name from another namespace). It's not like it's much of a bother because Eclipse has Ctrl + Shift + O to automatically add/remove/re-order the import according to android rules. Not to mention, everywhere else in android, explicit imports are used.

Patch-set: 1