Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

I think icons in folders should have labels.
Now we have a clean homescreen and that was the reason why I am using it in adw. 
I think labels in folders are good. So you can have bookmarks in folders and know what site is marked.
But it's only my opinion.

Patch-set: 2