Update patch set 3
Patch Set 3:

is it possible to change the position of the icon by hiding the lable?
if you have a 1x1 widget next to an icon without lable it looks a bit strange.
so is it possible to center the icon in the place for icon and its lable in the hight?

Patch-set: 3