Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Strange effects now. It seems to switch from screen 1 to 5 and from 5 to 1. But then you are in a set of 5 empty screens! When you then scroll back to the "real" 5 screens with icons then the icons are sometimes slightly moved to the left by some pixels. Needs more rework.

Tested on htc saga...

Patch-set: 2