Update patch set 5
Patch Set 5:

Not really sure which specific commit that broke the scaling on the maguro but comparing the screenshots below the difference is clear.

Stock-like: http://i.imgur.com/18haj.png

Now: http://i.imgur.com/Ms8ZR.png

As you can see if affects everything from the widget size (looks quite bad) to the dock. The homescreen and widget scaling can be adjusted to look more like the first image by setting a horizontal and vertical padding of 2 or 3. The height of the dock divider etc. won't be corrected though.

Patch-set: 5