Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

Fails in a similar way as Patch Set 2 on my maguro

Going from the rightmost screen further to the right, shows me an endless supply of empty homescreens, instead of switching to the leftmost screen.

Going back from there now shows the normal homescreens exactly shifted by 50% of the display width, i.e. the left half of the screen only shows the background picture while the right half of the screen shows the left half of the homescreen content.

Ah. The horizontal shift appears to be related to how far I scrolled into the "uncharted territory" before swiping back, it is not always 50%.

Futhermore, flipping from portrait to landscape FIXES the shift, and the fix remains when flipping back to portrait.

Hope this helps understand the issue!

Patch-set: 3