Update patch set 5
Patch Set 5:

nebkat: I still thought we wanted to keep the stock behaviour as the default? As it is now it's impossible to get it to look like stock. At the very least I think the padding should be something other than 0 as it breaks the look of widgets.

Patch-set: 5