Update patch set 2
Patch Set 2: (3 inline comments)

I've made some comments again. For me it's really complicated to translate a "standalone" English word to German. Sometimes it is nearly not possible to translate from English to German :-( 

I have the same problem with "deep idle" for example. It is not complicated to translate it, but do have the same sense...?! or "touch-key"

hopefully we can find a good solution...

Patch-set: 2