Update patch set 7
Patch Set 7: I would prefer that you didn't submit this

After the new "doctrine" to not change that which has a msgid... I would be forced to again change back the translation for Homescreen (Seite), to the established usage, which is .... tada.... Startbildschirm and Startseite, in one or the other place.

So... What should I do?

Patch-set: 7