Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

This looks horrible for the standard transition effect in my opinion.

Also, this change removes the visible indication that you have hit the end and therefore in my mind goes against the Android design. This makes it feel like iOS, not Android.

Patch-set: 1