Trebuchet: Add overlayable config for tablet workspace grid size
Tablets with a 4x3 screen such as the HP Touchpad do not
work well with automatic grid size setting.

This creates a flag in config that can be overlayed at
build time to allow user settings to adjust grid size.

The maximum settable row and column counts are also overlayable.

The default behavior is unchanged.

Change-Id: I0577f54aad4e247277483378b413052ea488fd48