Update patch set 1
Patch Set 1:

@Robert
I abandoned 23798 since it broke the animation 'scroll to screen of added widget' animation
AFAIK, adding widgets to non displayed screens is feature of cm.
i havent seen the supporting code for it in aosp.
I think that here is the right place to fix it since the boolean "willSnapPage" decides whether to animate or not to the screen. the f/c happens when it tries to animate to not displayed screen.
Adding the widget to not displayed screen(ex. screen no. 3 when user picked 2 screens) should be possible if first screens are full, but it shouldn't try animate to it.

Patch-set: 1