Update patch set 1
Patch Set 1:

@Idan -

No need to be sorry, glad you jumped in and supplied a patch.  FYI, your patch very well could have been fine, but I think we wanted to catch sooner in the flow of things.  Thanks for supplying the patch!

I added you as a reviewer to my supplied patch, if you could cherry-pick it and put it through additional testing that would be great.

Patch-set: 1