Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

As I said on irc before,I would prefer, if the launcher button can be placed by drag and drop or any other solution that does not require an option.

Remember: we try to keep cm free of too many options - especially the ones that can be avoided. Additionally it's more user friendly to change the position on the fly without changing to options dialogue.

Sorry I forgot to mention that before

Patch-set: 2