Update patch set 5
Patch Set 5:

Chad: I didn't mean it failed to build. What I was trying to say was that I didn't have time to get it built. I had a feeling my wording might not be 100% clear though.

To clarify for anyone else reviewing this, the drawable is shown when you hold an icon/widget and move it to the next homescreen.

Patch-set: 5