Trebuchet: Add overlayable config for tablet workspace grid size
Tablets with a 4x3 screen such as the HP Touchpad do not
work well with automatic grid size setting. This problem is described in
CyanogenMod Issue 5798: Trebuchet Tablet. This fixes that issue.

This creates a flag in config that can be overlayed at
build time to allow user settings to adjust grid size.

The maximum settable row and column counts are also overlayable.

This is a port of: http://review.cyanogenmod.org/#/c/21067/
which was merged into jellybean.

The default behavior is unchanged.

Change-Id: I3d7cf5e121a423a0033310eee88e498bc080af3c