Update patch set 8
Patch Set 8: Looks good to me, but someone else must approve

Well, after an exhaustive examination, a compilation and a complete test, I think the translation is ready. Sorry for the high number of patchset.

I have taken the liberty of adapting the translations to the real action, instead of doing a literal translation. Thus, it is clearer for the Spanish-speaking what is going to happen to enable the option.

Merry Christmas.

Patch-set: 8