Update patch set 3
Patch Set 3:

This is a big step backwards.    If this was added just to provide a way to add "Actions" i highly sugget we find a better way.

One example is bring back the old wallpaper launcher and add an icon to take you to a new tab in the app/widget chooser called 'Actions" .

Another option is to not allow the last "Apps" action to be deleted.  At least one is always required.

Either option is far better than bringing back a long press menu to add apps and icons and to have to go through a giant list.   

Huge regression IMHO and it is apparent to others as well.

Patch-set: 3