Update patch set 1
Patch Set 1:

That's what I did initially, but it also gave me a path error locally - and it somehow resulted in my local change suddenly not belonging to any local branch (checked with "git status"). I have no idea, how this happened - I haven't encountered an issue like this before. Not being sure how to fix this mess, I figured it was probably best to create a new local branch and resubmit it and abandon the old one.

Patch-set: 1