Update patch set 3
Patch Set 3:

Seriously, that not-even-funny commit is going to stay in Trebuchet? I hope you're kidding as it represents everything that CM is against - like extraneous "settings" "just because it can be done". I don't think it should be translated - even if we do, noone is going to understand it.

Patch-set: 3