Update patch set 4
Patch Set 4:

So, what you're saying is, other developers don't know how to have fun? Other devs don't know how to let off a little steam? No other project in the history of software development has ever had a particle joke in their software? You must be used to Oracle and Apple software.....

Welcome to a group of developers that knows how to have a little fun. That stick.... pull it out of the mud.

Patch-set: 4