Update patch set 1
Patch Set 1:

I think I wasn't clear on one thing - I'm not against easter eggs, they are quite delightful, but this change is more like an intrusive nonsense. What do you think, how many of the users will actually get it?

This revert wouln't be here if it was just a line of "Y u up at 3AM?!", that would be funny, but a lot of Koush is not funny, it's ridiculous.

I'll post a patchset with reverted translations once I have my PC back if you wish but I'd love to see some judgement on this one until then.

Patch-set: 1