Update patch set 3
Patch Set 3: I would prefer that you didn't submit this

Please don't remove the msgid of a string. This makes it easier to import AOSP translation updates.
Also, please don't include config_* strings; they are to be used in device overlays for a special configuration, not to be translated.

Patch-set: 3