Update patch set 1
Patch Set 1:

There's this awesome bug:

1. Turn expanded desktop off, and reboot your phone.
2. Now hold a hotseat item for a sec, until it vibrates, and drop it in that same place.
3. Enjoy floating icons that cannot be moved.

It also gives wrong calculations about workspace size, which ends up in cluttered icons on "all apps" menu.
This fixes that, but I can't fix the first bug mentioned, so, looking for help here. A temporal workaround is to kill launcher via activity manager on PhoneWindowManager (so it works with all launchers, not only trebuchet), but the best solution would be to fetch again all values when screen sizee changes.

Patch-set: 1