Update patch set 3
Patch Set 3:

Maybe split the padding definitions into external dimension files and use the original large paddings in the dimen-w340dp bucket (~340 should be enough to fit the tabs on any language, the icons and some extra space)? That won't fix the problem with antiquated smaller resolution screens, but personally I can't think of an easy solution for that. I'm sure the UI/UX team would agree that the current icons look a bit off.


BTW, patch set 3 seems to break this on my i9000 - the overflow no longer shows up (I have the override enabled in the settings).

Patch-set: 3