Launcher: Show market icon in menu if no space
With the old AOSP paddings the icons look cleanly separated and less
crammed together. In addition, they follow the 48dp touch target
guideline.

Depending on the user's language and screen size/resolution the app &
widget tabs may be too wide for both of the market and overflow buttons
to fit on-screen. To account for cases like these, some code has been
added to overflow the market button to the menu.

Change-Id: I2e7702034151bb6dd6eac5cc42cf4568013f1b7f