Update patch set 3
Patch Set 3:

Jorge: that shouldn't be happening. I can't reproduce this on the i9000 with Spanish and overflow enabled: does the i9100 have the same 240 density? The buttons are clearly overflowing and the market should be getting hidden.

What's the idea with the overflow visibility code being changed constantly? Right now the code doesn't follow what the comments say (preferencesVisible is actually "preferencesInvisible"). Should it be visible in CM or not?

Patch-set: 3