Update patch set 5
Patch Set 5:

Jorge/Nebkat,  we need to update this a bit.  Firstly,  the cLock widget default size needs to be specified to be full width. That way it will be centered. Also,  suggest we remove the power widget from the 2nd screen.  We have quick settings so it's redundant.

Patch-set: 5