Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

nebkat, i think that there is not space to play with the padding. In portrait, there is a bit of space to increase the hotseat, except if strech screens is set. And now, this is the default option (steve merged a patch some days ago).
And in landscape, shortcuts are beside each others with no too much space. And then it needs to adjust the side of the items properly for landscape and portrait modes.
I'll try to play with the paddings and sizes, but imo the scale of the icons is the best solution.

Did you play with strech screens and landscape mode?

Patch-set: 2