Update patch set 1
Patch Set 1:

Do we even want this as an option?

It's actually broken with it *off* on some devices such as the Galaxy Note. I haven't looked into why, but icons will actually be clipped from the screen edges with the option disabled. For devices where it's working, it looks terrible.

Patch-set: 1