Add handling for reverse default rotation in Trebuchet
Example of a device which needs the config_reverseDefaultRotation
and needs this patch for rotation to function as expected:

Kindle Fire HD's have a camera on the "left" side of the device when
holding it upright in portrait mode.  The power buttons and headphone
jack are at the "top".  For a launcher to correctly select the
default landscape mode, the rotation should be reversed and rotate
counter-clockwise so that the camera side is now at the "top".

Change-Id: Id3e5d2ee391700ab6772293808ba442c78446a9f