Update patch set 5
Patch Set 5:

Jorge,

I am well aware of our stance on icon packs. I only proceeded with publishing this patchset once I got a green light from nebkat/arcee.

Reason being that although it would be awesome for something like this to be done at the theme manager level, currently there is no support for it and the modifications required to add support for it won't be happening any time soon.

Patch-set: 5