Update patch set 9
Patch Set 9:

I reject this rejection of icon packs.

There is a huge community of people creating and using them. Not taking part in this is just plain idiotic and defies logic.

Yeah, we should try to do it right, but not supporting them just costs us users who are happy to switch to random closed source ad supported apps to get this feature.

We need this ASAP.

Patch-set: 9