Update patch set 10
Patch Set 10:

Yea so the second iteration will be ability to choose icons from icon packs and assign them individually to shortcuts.
Currently pressing the icon on the dialog does nothing.

I don't want to rely solely on icon pack's ability to provide PICK_ICON_PACK, so i'll just have to manually populate a list of icons from it. Not to mention even if an icon pack supports PICK_ICON_PACK, it just returns a bitmap as a result and not the resource name which is what id rather have.

Patch-set: 10