Update patch set 5
Patch Set 5:

Just try it on my Galaxy Nexus and notice something strange.
I installed Nox icon pack.
When i try to change icon of an apps on Desktop all is good.
But when i try to change icon of an apps which are in a folder the icon doesn't change directly, i add to return in Edit menu and just click OK a second time to apply the new icon.

Patch-set: 5