Update patch set 6
Patch Set 6:

With default icon, if you go to the desktop and choose to edit an icon and apply it.
If you next go to Trebuchet option and choose the Default icon pack this icons won't change.
They need to be manually edit to Default theme before try to apply another icon pack

Patch-set: 6