Update patch set 7
Patch Set 7:

I'm not sure that is a sound solution, particularly in the invent of multiple packs being installed at once: would we consume all the notifications, make them all interactive? 

Currently, the functionality in this patch (in the event of multiple detection) offers up the apply action for the last installed (newest) pack.

Patch-set: 7