Update patch set 1
Patch Set 1:

https://lh3.googleusercontent.com/-b63P1RiplM4/Ulb8Sw5Lh0I/AAAAAAAAAhc/4SiYn64Ako8/w765-h478-no/Screenshot_2013-10-10-15-12-22.png

Heres a screen shot of the landscape bug. I narrowed down to when it occurs. It has to do with the landscape orientation being loaded while in the app drawer. It seems like it loads the margins for the app drawer and also applies them to the homescreens. It happens when backing out of an app into the landscape drawer or if you rotate to landscape while in the app drawer.

To force the issue, open the homescreen in portrait, then open the app drawer, rotate the screen to landscape and then exit the app drawer, scroll back and forth so it loads the other pages, you'll see that the pages are too close and end up under the dock and search bar.

Patch-set: 1