Update patch set 1
Patch Set 1: -Code-Review

nebkat, nothing personal ... But this kind of thing is sad. It's not the first time, and probably will not be the last ... Sometimes we work to solve a bug and someone who have merge access (not talking about one specific repo) appears with the solution and push directly (or rebase the tree completely), without pass through review. It's not easy to mix private and public work...
Thanks for all your hard work on Trebuchet, it's the best Launcher.

Patch-set: 1
Label: Code-Review=0