Update patch set 1
Patch Set 1:

Sorry if that is how it looks. I'll admit there were times when I had things done but kept them private for too long until someone else redid the feature, but this was mainly due to my (sometimes busy) schedule and development cycle (I kept things private until there was a reasonable amount of features to push). However in this case I honestly must've forgot to push the proguard fix. I was inactive during the past week because of holidays, seeing friends, etc and at the moment I'm moving things around the house including the PC which isn't even plugged in. As for the Gerrit vs. direct push you should really only use GitHub as a reference of what patches have been merged for any repo.

If it isn't too much to ask I think it would be best if people could contact me on IRC, Gmail or G+ before working on patches that they woukd mind getting discarded in case I have them WIP locally. It could also be the case that I would like to do a certain implementation myself - I hate having to reject a patch that someone has spent a lot of time on because the code/implementation isn't up to standard.

Patch-set: 1