Update patch set 2
Patch Set 2:

I'd really like to avoid the excessive range of possibilities this feature offers. I'd rather not offer an option to the user than renders their homescreen in an unusable state (for eg, labels getting cutoff...etc). On top of that, the extreme settings break the expectations of widgets, since they expect specific amount of padding between cells for n number of cells.

I have added David to this conversation, as we discussed the issues with this feature previously.

I'd much rather have something similar to what Steve mentioned earlier to me, which is similar to what gmail does  and that is predefined configs.

For eg : Comfortable	Cozy	Compact

This allows us to be in control of the settings and at the same time allows users to have this flexibility.

Thoughts ?

Patch-set: 2
Reviewer: Danesh M <940@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0