Update patch set 1
Patch Set 1:

While we are addressing the default screen, I have to ask why you've chose the current implementation. 

I feel like this would be so much cleaner if it was a simple pin icon on the top left corner of the overlay view for each page. User would simple tap that pin to set it as default. 

Quick and lazy gimp edit http://imgur.com/AQbvCeX

Patch-set: 1