Update patch set 1
Patch Set 1:

Yeah, the problem is that Trebuchet is killed when a property changed to refresh all its internal vars. So when you switch from the recents panel to the trebuchet settings there are a valid instance (the previous was killed) and the system restore the settings (the headers one) and set the default view to the first one.

Patch-set: 1