Update patch set 1
Patch Set 1:

What is the problem with the rename?
I could not get the HiddenAppsActivity to be shown without really renaming the package.
Code in preferences_drawer.xml:
        <Preference android:key="ui_drawer_hidden_apps"
            android:title="@string/preferences_interface_drawer_hidden_apps_title"
            android:summary="@string/preferences_interface_drawer_hidden_apps_summary">
            <intent
                android:targetPackage="com.cyanogenmod.trebuchet"
                android:targetClass="com.cyanogenmod.trebuchet.settings.HiddenAppsActivity"/>
        </Preference>

This does not seem to work, if you rename the package with aapt command. It always causes an activity not found exception opening HiddenAppsActivity. No matter which package you set there.

There is a bug anyway, got to recheck this.

Patch-set: 1