Update patch set 3
Patch Set 3:

If you hide the labels, you have more space on the screen. Most likely you have enough space for another row of icons, so this adds this row, if there is enough space.
The widget size is a problem though, but I think, as changing the number of rows is already planned, we can discuss how widgets should be handled with changing cell sizes.

Patch-set: 3