Update patch set 2
Patch Set 2:

apologies for the whitespace issues... i have vim setup for a default indent of 4 for xml & java - is it 2 for cm? i could find google's coding style guidelines, but not cm's (i'm assuming you just use google's?).

regarding the number-picker :), like we discussed:
i both understand and agree with the "don't give the user unusable options".
but... i think it's a good compromise with having a limit (like predefined options would do), while keeping the famous CM flexibility, which predefined options would not (which is important to me, and the reason, i personally use it) - without giving the user the option to render the homescreen unusable.

Patch-set: 2