Update patch set 2
Patch Set 2:

I like the user adjustability. I don't think that the user should be able to set unuseable grid sizes. If we make this enabled by an overlay is set by the device maintainer along with default and maximum grid sizes for that device. This way, the user won't have the option to choose unreasonable grid sizes. Make it only enabled if the device maintainer has set maximum row and column sizes.

Patch-set: 2