Update patch set 2
Patch Set 2:

David - that's the difference between this suggested patch and previous ones. 
even if the device maintainer does not provide an overlay with a reasonable maximum value, a REASONABLE maximum value is calculated at runtime to make sure the user cannot set an unreasonable value (bullet #3 in the commit message).
this behaviour is enabled by default, so device maintainers do not have to add anything for it to work, but it is overridable for edge cases or maintainers who think values should be different.

Patch-set: 2