Update patch set 5
Patch Set 5: Code-Review-1

Right, messed that up. Wanted to write "Reduce row size"
However this is not complete and contains some bugs:
The dynamic grid is evil... It should only be used to provide default values for the row and col count. The real number of rows/cols should be set in the constructors. Furthermore there are different grids scattered around. (e.g. occupied in LauncherModel and CellLayout)
There should only be one grid, that is either managed by the screens and LauncherModel accesses it from there, or the other way round: The LauncherModel contains all grids for all screens and the screens access this.
Currently the grids may become inconsistent if you forget to update all.
So I would rather rework the Dynamic grid system (or better: its usage) first, before doing something like this.

Patch-set: 5