Update patch set 3
Patch Set 3:

I'm experiencing a tiny bug on the latest nightlies. On reboot, my wallpaper is offset to the side and I have to swipe screens for it to return back to normal, centered. I thought it might be the recent wallpaper commits but I installed Google Now launcher and the issue isn't present there so not wallpaper related. My guess is it might be this commit but haven't been able to test that theory. I'll run a build tomorrow when I can to test.

Patch-set: 3
Reviewer: Ricky Salinas <5505@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0