Update patch set 2
Patch Set 2:

(1 comment)

Totally agree with you. I like the solution.
Just a few things: Why not use the double-number-picker from CM 10? Don't know how this looks like, but IMO a single entry for grid size that sets rows AND columns is better than 2.

With one of my patches it is possible, to hide the labels of the icons. This results in more space, hence possibly more rows. And you can set that option for home screen and app drawer separately. It would be great, if you could set the grid size for home screen and app drawer also separately and (at least later) use also the free space from the labels to calculate a maximum. Maybe even couple this 2 options: If row count is at the last stage, hide the labels automatically.
But first step would be different grid sizes for home screen and app drawer (someone already suggested having different grid sizes per home screen...) which is a bit difficult with the dynamic grid used as it is now.

Patch-set: 2