Update patch set 2
Patch Set 2:

This feature is going to evolve past just cleaning the all apps drawer in May with a UX overhaul for Trebuchet. 

But even with that, the feature is 'hidden apps', not 'do not show app in drawer' (filter). Filtering an app from a view is not "hiding". 

Like I said, I want this to go deeper - hidden apps should be securely accessible, but at the same time, hidden means 100% hidden, not hidden in just one factor of the word. 

Back on topic, the notion that by hiding an app the user doesn't expect the homescreen to change - this is a failure in the current UX model, not with how this feature was integrated.

Patch-set: 2