Update patch set 2
Patch Set 2:

Okay, I get what you are saying, but when you say "the feature is 'hidden apps', not 'do not show app in drawer' (filter)" ...and then take a look at the summary, it clearly says "Hide apps from the drawer".

People expect it to do just that, hide/filter app icons from the drawer only.

Whether this will change in the near future is not important right now and I still think this little patch is relevant.

But it's up to you.

Patch-set: 2