Update patch set 2
Patch Set 2:

Although this is already merged, just my opinion:
I really wanted to have this option to hide apps from the drawer, nothing more. This is why it is (again!) in the apps drawer section. And in this section "hidden" means "hidden" (from the drawer), so actually the checkboxes should be removed and home screen not changed. But I understand, that some people may want to hide the apps from "everywhere". This is what the checkboxes are intended for.
Why would you want to hide apps but not uninstall them? NSFW content?

Bottom line: If you implement a "hide apps from everything" function, leave this to hide apps from the drawer. As you see, I'm not the only one needing it.

Patch-set: 2
Reviewer: Alex Grund <8616@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0