Update patch set 1
Patch Set 1:

The Remove and Info buttons are actually accessible in the top right corner. See http://imgur.com/a/SXOk7#2

I don't know how the drop target bar is supposed to look in landscape and how it looks on devices where is_large_tablet is true. I just took a simple approach and changed the LinearLayout orientation to vertical if the device is in landscape.

Two smaller issues I have so far. The Remove and Info icons are further to the right than the icons of the search bar and the transition animation does not look that nice because the icons are aligned differently. The search icons are at the top while the others are stretched over the whole height and centered.

Patch-set: 1
Reviewer: Thomas Wendt <5678@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0