Update patch set 6
Patch Set 6: Code-Review-1

Nexus 4:
Combining the large icons withe the the icon labels on and more then 5 columns creats a bug:
a part of the lable is cut on the buttom.
a part of the icon is cut on the top.

Sometimes i can select 6 rows and columns and on others i can  select 5.

changing the grit size to a smaler number, will hide the apps or widgets on the 5. or 6. Row/collumn. increasing the grid again brins the Icons and widgets back.
they dont shift over to the next free Space on an other page.(not tested with sizes smaler then 4!)

have tested 5x5 grid with some widgets. no bugs or glitches.
everything refits perfectly.

Patch-set: 6