Update patch set 6
Patch Set 6: Code-Review+1

>>Nexus 4: Combining the large icons withe the the icon labels on and more then 5 columns creats a bug: a part of the lable is cut on the buttom. a part of the icon is cut on the top.


this used to happen in Launcher3 before the change to Trebuchet. only Launcher3 had no option to reduce the icon sizes so you had to make do with the ugly clipping. i did, 5x5 was more important than looks to me.

avoiding clipping is difficult to do right. if u have large icons and labels, the max no-clipping grid size is going to be 4x4. should the UI reflect that? that's tricky: let's say it limits you to 4x4, fine. then you go and reduce icon size (the user has to know this! (usability issue)), get back and set 5x5, fine. then you go back to icon size, and now the large icon option has to be disabled! not only that, the user has to know why it is disabled (or you have another usability issue). and interaction is 3 ways, because you have to include the label setting in the mix too. this is unnecessarily complicated to code, and the result is not user friendly. i vote AGAINST going in this direction.

a better alternative IMHO: you set the grid size freely and the label toggle freely. (btw, this is great to move launcher settings among devices, something i always do. i change devices all the time.) the system then calculates the max icon and text size according to these settings. the large icons option is gone. in its place there is an icon size setting, going from 100% to say 50% of the calculated max current size. this option should scale down icon and labels to add padding between the icons. this way all settings are independent and the UI is simple to code and predictable to use. maybe the icon size setting should be discreet: a normal setting to replicate the padding on standard 4x4 Trebuchet today (the default) and higher and lower discreet settings.

also the label on/off toggle could be replaced with discreet options for labels: normal, smaller, smallest, hidden. this would let you choose smaller text (with respect to icon size) to fit longer labels. the text in this case would be scaled by the product of TWO settings: label type and icon size.

----------------

>>changing the grit size to a smaler number, will hide the apps or widgets on the 5. or 6. Row/collumn. increasing the grid again brins the Icons and widgets back. they dont shift over to the next free Space on an other page.(not tested with sizes smaler then 4!)


holy cow! please dont implement any icon/widget shifting around!! home screen management is a tedious operation and i dont want a silly seemingly innocuous settings change that i maybe did out of curiosity at some point (ie: reducing the grid size) to utterly destroy my home screens, all damn 8 of them!

i understand the reluctance of accepting the current behavior (inherited from Launcher3) of icons and widgets disappearing on grid size reduction (and later reappearing on an increase), but distroying my home screen is *NOT* a solution, it is making it worse.

IMHO this is easy to code and good enough: find the number of empty rightmost columns and empty bottommost rows that are available in ALL the homes, then use this to set a minimum grid size. (caveat: gracefully handling the corner case of fully empty home screens.) the UI row count spinner screen can include a text saying "The minimum configurable number of rows is currently 4. To reduce the number of rows further, please empty bottom rows in all of your home screens." same goes for the column UI spinner. also, on startup Trebuchet could run this min grid test and increase the grid size if necessary. this is to recover from broken configuration, or to seamlessly switching from a 5x5 Launcher3 to Trebuchet. (right now the 'upgrade' looses icons and widgets.)

----------------

finally, i think settings for home and drawer should be independent. the sharing of icon size settings does not make much sense to me and spreads the home screen icon appearance setting across different screens (low usability). the text font style could be shared if needed. the drawer appearance customization is IMHO much less important than home screen's. the drawer manages its own content while the home is fully manual.

----------------

i am voting this +1 as it is, because it is better to have non-perfect functionality now than promise of perfection in the future. this patchset works: the problem with clipping is trivial, and icons disappearing, well its ugly, but they reappear if you increase the size back.

at least it lets you recover your old Launcher3 icons if you had them. people already suffered from this icon loss on L3->Tre upgrade, and i bet there are many phones out there with 'phantom' icons. what happens if the user wants to place a new icon on a space spanned by a phantom 3x3 widget? dont have a clue! :)

Patch-set: 6