Update patch set 4
Patch Set 4:

Addressed issue with Drawer/HomeScreen settings cross over.

I believe the issues with the mdpi and hdpi devices is due the lack of transparency support for the status bar and nav bar in those devices. Fix for this will get rolled with any modifications required of the UX as suggested by nebkat.

Patch-set: 4