Update patch set 9
Patch Set 9:

In Nexus 4, the Wallpaper text for other languages wraps to 2 lines and it does not fit.


Screenshots:

https://lh4.googleusercontent.com/-HA8v4se6hIY/U4SNan744iI/AAAAAAAADVw/SWThLJOgt1Y/w597-h995-no/


https://lh5.googleusercontent.com/-uXwcU2-RFbU/U4SNapG9B1I/AAAAAAAADVo/CbM4svm7fAs/w597-h995-no/



Some more feedback: 

Would it be possible to make the background of the settings transparant like the original? It looks really ugly now imo and a transparant panel would make it look 1000x better.

Also, the animation is much slower than it was before, also much slower than Google Launcher settings. I think because it looks like a two-step process: first the workspace shrinks and then the menu panels appear.

If transparancy is applied, you will not notice this as much, speeding things up


Maybe when you slide the panel up, transarancy can be made less/background darker to make things more visible.

Patch-set: 9