Update patch set 9
Patch Set 9:

I agree with Michael. ON/OFF Labels don't look good in other languages. I vote, again, for checkboxes....


There is another issue introduced with this patch: the margins around the edges are bigger, meaning that text labels that used to fit in the old situation, don't fit in the new UI. 

See this screenshot to compare: https://www.dropbox.com/s/p5z824fo7f0h4ze/Screenshot_2014-05-27-19-40-15.jpg

The word "Achtergronden" used to fit and now it does not fit anymore, it wraps to a second line. It is an AOSP string, so we cannot change it either.



I hope you also agree that this looks worse than before, because of the lack of transparancy.

Patch-set: 9