Update patch set 1
Patch Set 1:

Here's my logic:

Inc builds off of CM mainline code branched as necessary. Introducing Google references as it pertains to these Inc elements makes sense, as its going to be what is on shipping hardware and is licensed. 

The circumstances that would allow for GCamera to exist on a default workspace are not attainable via the Inc licensing (ie not CDD compliant). So the only way this would even be triggered in a default config is by the community distributing their own pack that disables Camera2 and replaces it with GCamera. This, though minor, makes the difference to say that it should not be in the default workspace.

Patch-set: 1