Update patch set 1
Patch Set 1:

I have attached a couple pngs that illustrate my following comments:

http://imgur.com/a/7BjCZ

In the redline file you can see the letters I am referencing.

A. The default screen icon is violating the clear space for the notification bar, I suggest moving the icon down to fit under the 25dp height of the bar

B. The screen preview is overlapping the default icon and cutting through it, I think we should scale the preview pane down so that there is a 4dp padding between the pane and the icon.

C. There appears to be an artifact or drop shadow remnant from the drawer panel that is now transparent. I suggest not removing the shadow but instead adding a small amount of opacity to the drawer panel. As it looks right now I don't feel as a user that it is something I can grab and drag up. There needs to be some amount of surface to grab. In my before and after png you can see what an 8% opacity did to still retain the transparent look but allow for it to still be recognized as a grabbable element.

D. The icons appear too low on the panel and are uncomfortably close to the on screen nav buttons. I have suggested a padding of 56dp from the bottom.

Patch-set: 1