Update patch set 1
Patch Set 1:

If we stay transparent then we need to remove the drop shadow I mentioned in point C.

As for the arrow, I really like the animation as it brings the panel to life in an unexpected way. I think having it point up in the beginning helps to let me know that I can pull it up. It already appeared from the bottom so I naturally think that I can pull it back down.

Patch-set: 1